# openapi.yaml
openapi: 3.1.0
info:
  title: Envelope Sync Relay API
  version: 1.3.0
  description: |
    Zero-knowledge document sync relay for offline-first applications.
    Full specification: docs/envelope-sync-relay/README.md

    **v1.3:** Application registry — `X-ESR-App-Id`, namespace–app binding, operator admin API,
    developer self-service portal, origin verification, pairing `allowedAppIds`.
    See docs/envelope-sync-relay/en/16-APP-REGISTRY.md

    **v1.2:** Multi-document per namespace (`/documents/{documentId}/...`).

    WebSocket notifications (v1.1): wss://{host}/v1/namespaces/{namespaceId}/notifications
    See 13-WEBSOCKET-NOTIFICATIONS.md — not covered by this OpenAPI file.
servers:
  - url: https://sync.example.com/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local development
...
components:
  securitySchemes:
    DeviceToken:
      type: http
      scheme: bearer
      bearerFormat: device_token
    AdminToken:
      type: http
      scheme: bearer
      bearerFormat: admin_api_token
    DeveloperToken:
      type: http
      scheme: bearer
      bearerFormat: developer_jwt
...
  schemas:
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    Limits:
      type: object
      properties:
        freeDeviceLimit:
          type: integer
        purchasedSlots:
          type: integer
        maxDevices:
          type: integer
        activeDevices:
          type: integer
    OnLimitReached:
      type: object
      properties:
        mode:
          type: string
          enum: [payment, block]
        slotPackages:
          type: array
          items:
            type: integer
...
    Device:
      type: object
      properties:
        deviceId:
          type: string
        clientDeviceId:
          type: string
        label:
          type: string
        pairedAt:
          type: string
          format: date-time
        lastSeenAt:
          type: string
          format: date-time
        isCurrent:
          type: boolean
    RateLimitQuota:
      type: object
      required: [action, limit, remaining, resetAfterSeconds, windowSeconds]
      properties:
        action:
          type: string
          enum: [global_ip, put_document, recover, pair_device, pairing_token]
          description: Internal action id; JSON `rateLimits` keys match this value (document PUT uses `put_document`).
        limit:
          type: integer
        remaining:
          type: integer
        resetAfterSeconds:
          type: integer
          description: Seconds until the oldest event in the sliding window expires.
        windowSeconds:
          type: integer
    RateLimits:
      type: object
      description: Optional on success when limits.rateLimit.enabled; keys are action ids.
      additionalProperties:
        $ref: '#/components/schemas/RateLimitQuota'
...
    OriginVerification:
      type: object
      properties:
        dnsHost:
          type: string
        dnsTxt:
          type: string
        wellKnownUrl:
          type: string
          format: uri
    AppOrigin:
      type: object
      properties:
        id:
          type: string
          format: uuid
        origin:
          type: string
          format: uri
        verifiedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        verification:
          $ref: '#/components/schemas/OriginVerification'
          nullable: true
    AppBundle:
      type: object
      properties:
        id:
          type: string
          format: uuid
        platform:
          type: string
          enum: [ios, android]
        bundleId:
          type: string
        verifiedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    AppSummary:
      type: object
      properties:
        appId:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [web, native]
        status:
          type: string
          enum: [pending, pending_verification, active, suspended, archived]
        originCount:
          type: integer
        bundleCount:
          type: integer
        namespaceCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AppDetail:
      allOf:
        - $ref: '#/components/schemas/AppSummary'
        - type: object
          properties:
            origins:
              type: array
              items:
                $ref: '#/components/schemas/AppOrigin'
            bundles:
              type: array
              items:
                $ref: '#/components/schemas/AppBundle'
            hasClientSecret:
              type: boolean
    PaginatedApps:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AppSummary'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...
    PairingTokenResult:
      type: object
      properties:
        code:
          type: string
          pattern: '^\\d{6}$'
        expiresAt:
          type: string
          format: date-time
        qrPayload:
          type: string
        allowedAppIds:
          type: array
          items:
            type: string
        rateLimits:
          $ref: '#/components/schemas/RateLimits'
  responses:
    Error400:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
    Error403:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
    Error429:
      description: |
        Rate limit exceeded (`RATE_LIMIT_EXCEEDED`). No top-level `rateLimits` on errors —
        use `error.details.rateLimit` (single quota). Matching `RateLimit-*` or prefixed
        headers (e.g. `RateLimit-PutDocument-*` for `put_document`).
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorBody'
              - type: object
                properties:
                  error:
                    type: object
                    properties:
                      details:
                        type: object
                        properties:
                          retryAfterSeconds:
                            type: integer
                          action:
                            type: string
                            enum: [global_ip, put_document, recover, pair_device, pairing_token]
                          rateLimit:
                            $ref: '#/components/schemas/RateLimitQuota'
paths:
...
  /namespaces/{namespaceId}/devices:
    get:
      tags: [Devices]
      summary: List devices
      security:
        - DeviceToken: []
      parameters:
        - name: namespaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  devices:
                    type: array
                    items:
                      $ref: '#/components/schemas/Device'
                  limits:
                    $ref: '#/components/schemas/Limits'
    post:
      tags: [Devices]
      summary: Redeem pairing code (new device)
      parameters:
        - name: namespaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pairingCode, deviceLabel, clientDeviceId]
              properties:
                pairingCode:
                  type: string
                deviceLabel:
                  type: string
                clientDeviceId:
                  type: string
      responses:
        '201':
          description: Paired (may include `rateLimits` with `global_ip` and `pair_device`)
        '400':
          $ref: '#/components/responses/Error400'
        '403':
          $ref: '#/components/responses/Error403'
        '429':
          $ref: '#/components/responses/Error429'
  /namespaces/{namespaceId}/devices/{deviceId}:
    delete:
      tags: [Devices]
      summary: Revoke device
      security:
        - DeviceToken: []
      parameters:
        - name: namespaceId
          in: path
          required: true
          schema:
            type: string
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Revoked
        '403':
          $ref: '#/components/responses/Error403'
  /namespaces/{namespaceId}/pairing-tokens:
    post:
      tags: [Devices]
      summary: Create pairing token
      security:
        - DeviceToken: []
      parameters:
        - name: namespaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ttlSeconds:
                  type: integer
                allowedAppIds:
                  type: array
                  maxItems: 10
                  items:
                    type: string
                    pattern: '^esr_app_[a-z0-9_]+$'
                  description: Optional — restrict pairing redeem to listed active apps
      responses:
        '201':
          description: Token created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PairingTokenResult'
        '403':
          $ref: '#/components/responses/Error403'
        '429':
          $ref: '#/components/responses/Error429'
...
  /namespaces/{namespaceId}/limits:
    get:
      tags: [Limits]
      summary: Get slot limits
      security:
        - DeviceToken: []
      parameters:
        - name: namespaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Limits'
                  - type: object
                    properties:
                      onLimitReached:
                        $ref: '#/components/schemas/OnLimitReached'
                      canAddDevice:
                        type: boolean
...
  /admin/apps:
    get:
      tags: [Applications, Admin]
      summary: List registered applications (operator)
      security:
        - AdminToken: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, pending_verification, active, suspended, archived]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Paginated apps
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedApps'
    post:
      tags: [Applications, Admin]
      summary: Create application (operator_managed)
      security:
        - AdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [appId, name, type]
              properties:
                appId:
                  type: string
                  pattern: '^esr_app_[a-z0-9_]+$'
                name:
                  type: string
                type:
                  type: string
                  enum: [web, native]
                status:
                  type: string
                  enum: [pending, pending_verification, active, suspended, archived]
                origins:
                  type: array
                  items:
                    type: string
                    format: uri
                bundleIds:
                  type: object
                  properties:
                    ios:
                      type: string
                    android:
                      type: string
                clientSecret:
                  type: string
                  minLength: 16
      responses:
        '201':
          description: App created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppDetail'
  /admin/apps/{appId}:
    get:
      tags: [Applications, Admin]
      summary: Get application detail
      security:
        - AdminToken: []
      parameters:
        - name: appId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: App detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppDetail'
    patch:
      tags: [Applications, Admin]
      summary: Update application name or status
      security:
        - AdminToken: []
      parameters:
        - name: appId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                status:
                  type: string
                  enum: [pending, pending_verification, active, suspended, archived]
      responses:
        '200':
          description: Updated app
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppDetail'
    delete:
      tags: [Applications, Admin]
      summary: Archive application
      security:
        - AdminToken: []
      parameters:
        - name: appId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Archived app
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppDetail'
  /admin/apps/{appId}/origins:
    post:
      tags: [Applications, Admin]
      summary: Add web origin
      security:
        - AdminToken: []
      parameters:
        - name: appId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [origin]
              properties:
                origin:
                  type: string
                  format: uri
                verified:
                  type: boolean
                  default: true
      responses:
        '201':
          description: Origin added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppDetail'
  /admin/apps/{appId}/origins/{originId}:
    delete:
      tags: [Applications, Admin]
      summary: Remove origin
      security:
        - AdminToken: []
      parameters:
        - name: appId
          in: path
          required: true
          schema:
            type: string
        - name: originId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Updated app
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppDetail'
  /admin/apps/{appId}/origins/{originId}/verify:
    post:
      tags: [Applications, Admin]
      summary: Run DNS TXT or HTTPS well-known verification
      security:
        - AdminToken: []
      parameters:
        - name: appId
          in: path
          required: true
          schema:
            type: string
        - name: originId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  verification:
                    type: object
                    properties:
                      method:
                        type: string
                        enum: [dns, https, localhost]
                      verifiedAt:
                        type: string
                        format: date-time
                  app:
                    $ref: '#/components/schemas/AppDetail'
        '422':
          description: Verification failed (APP_ORIGIN_VERIFICATION_FAILED)
  /admin/apps/{appId}/bundles:
    post:
      tags: [Applications, Admin]
      summary: Add native bundle
      security:
        - AdminToken: []
      parameters:
        - name: appId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [platform, bundleId]
              properties:
                platform:
                  type: string
                  enum: [ios, android]
                bundleId:
                  type: string
                verified:
                  type: boolean
      responses:
        '201':
          description: Bundle added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppDetail'
  /admin/apps/{appId}/bundles/{bundleId}/approve:
    post:
      tags: [Applications, Admin]
      summary: Approve pending native bundle
      security:
        - AdminToken: []
      parameters:
        - name: appId
          in: path
          required: true
          schema:
            type: string
        - name: bundleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Bundle approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppDetail'
...